feat: Tag, Post, PostTag 엔티티 간 연관관계 매핑 (#20)
* feat: 신고 카테고리 항목 추가

* Update PULL_REQUEST_TEMPLATE.md

* feat: filetype enum값 추가

* feat: entity에 검증 어노테이션 추가 (#11)

* feat: entity에 검증 어노테이션 추가

* feat: entity에 검증 어노테이션 추가

* style: 업로드 파일명, 서버 관리 파일명의 필드명 변경

---------


* chore: 패키지 구조 개선

* Revert "chore: 패키지 구조 개선"

This reverts commit f27a0d95080e705a953d94cfb7ef365a6a1f49cb.

* chore: 패키지 구조 개선

* feat: Tag, Post, PostTag 엔티티 간 연관관계 매핑

---------
