+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,13 +1,3 @@
-[PR 제목은 커밋 메세지와 동일한 형식으로 작성]
-<!--- e.g. "[feat] 투표 기능 구현" --->
-
-<!---
-# 뒤에 머지 후 close할 이슈번호를 작성
-# 자동으로 close 됩니다.
-<strong>
-Closes #
-</strong>
---->
 ## 💡 작업 내용
 - [ ] 작업 내용
 
@@ -28,3 +18,5 @@ Closes #
 - [ ] 작업 도중 문서 수정이 필요한 경우 잘 수정했나요?
 - [ ] 테스트는 잘 통과했나요?
 - [ ] 불필요한 코드는 제거했나요?
+
+closes #이슈번호
