--- src/main/java/balancetalk/domain/ViewStatus.java
@@ -1,5 +1,0 @@
-package balancetalk.domain;
-
-public enum ViewStatus {
-    NORMAL, BLIND
-}
