--- src/main/java/balancetalk/domain/file/entity/File.java
@@ -1,37 +1,0 @@
-package balancetalk.domain.file.entity;
-
-import balancetalk.domain.Notice;
-import balancetalk.domain.file.enums.FileType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class File {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "file_id")
-    private Long id;
-
-    private String name;
-    private String path;
-
-    @Enumerated(value = EnumType.STRING)
-    private FileType type;
-
-    private Long size; // TODO 수정 필요
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "notice_id")
-    private Notice notice;
-}
