--- src/main/java/balancetalk/domain/post/enums/PostCategory.java
@@ -1,5 +1,0 @@
-package balancetalk.domain.post.enums;
-
-public enum PostCategory {
-    CASUAL, DISCUSSION
-}
