--- src/main/java/balancetalk/domain/report/entity/Report.java
@@ -1,45 +1,0 @@
-package balancetalk.domain.report.entity;
-
-import balancetalk.domain.report.enums.ReportCategory;
-import balancetalk.domain.comment.entity.Comment;
-import balancetalk.domain.member.entity.Member;
-import balancetalk.domain.post.entity.Post;
-import balancetalk.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Report extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "report_id")
-    private Long id;
-
-    private String content;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member reporter; // 신고한 회원
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
-
-    @Enumerated(value = EnumType.STRING)
-    private ReportCategory category;
-}
