--- src/main/java/balancetalk/domain/vote/entity/Vote.java
@@ -1,32 +1,0 @@
-package balancetalk.domain.vote.entity;
-
-import balancetalk.domain.member.entity.Member;
-import balancetalk.domain.post.entity.BalanceOption;
-import balancetalk.global.common.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Vote extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue
-    @Column(name = "vote_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "balance_option_id")
-    private BalanceOption balanceOption;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-}
