+++ src/main/java/balancetalk/module/Notice.java
@@ -1,0 +1,46 @@
+package balancetalk.module;
+
+import balancetalk.module.file.domain.File;
+import balancetalk.module.member.domain.Member;
+import balancetalk.global.common.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Notice extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "notice_id")
+    private Long id;
+
+    @NotBlank
+    @Size(max = 50)
+    private String title;
+
+    @NotNull
+    @Size(max = 2000)
+    private String content;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @OneToMany(mappedBy = "notice")
+    private List<File> files = new ArrayList<>();
+}
