+++ src/main/java/balancetalk/module/ViewStatus.java
@@ -1,0 +1,5 @@
+package balancetalk.module;
+
+public enum ViewStatus {
+    NORMAL, BLIND
+}
