+++ src/main/java/balancetalk/module/comment/domain/Comment.java
@@ -1,0 +1,57 @@
+package balancetalk.module.comment.domain;
+
+import balancetalk.module.report.domain.Report;
+import balancetalk.module.ViewStatus;
+import balancetalk.module.member.domain.Member;
+import balancetalk.module.post.domain.Post;
+import balancetalk.global.common.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Comment extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "comment_id")
+    private Long id;
+
+    @NotBlank
+    @Size(max = 100)
+    private String content;
+
+    @Enumerated(value = EnumType.STRING)
+    @NotNull
+    private ViewStatus viewStatus;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @OneToMany(mappedBy = "comment")
+    private List<CommentLike> likes = new ArrayList<>();
+
+    @OneToMany(mappedBy = "comment")
+    private List<Report> reports = new ArrayList<>();
+}
