+++ src/main/java/balancetalk/module/member/domain/Member.java
@@ -1,0 +1,79 @@
+package balancetalk.module.member.domain;
+
+import balancetalk.module.post.domain.Bookmark;
+import balancetalk.module.comment.domain.Comment;
+import balancetalk.module.comment.domain.CommentLike;
+import balancetalk.module.Notice;
+import balancetalk.module.post.domain.Post;
+import balancetalk.module.post.domain.PostLike;
+import balancetalk.module.report.domain.Report;
+import balancetalk.module.vote.domain.Vote;
+import balancetalk.global.common.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.OneToMany;
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.validation.constraints.*;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Member extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "member_id")
+    private Long id;
+
+    @NotNull
+    @Size(min = 2, max = 10)
+    private String nickname;
+
+    @NotNull
+    @Size(max = 30)
+    @Email(regexp = "^[a-zA-Z0-9._%+-]{1,20}@[a-zA-Z0-9.-]{1,10}\\.[a-zA-Z]{2,}$")
+    private String email;
+
+    @NotNull
+    @Size(min = 10, max = 20)
+    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-zA-Z])$")
+    private String password;
+
+    @Enumerated(value = EnumType.STRING)
+    @NotNull
+    private Role role;
+
+    @Size(min = 15)
+    private String ip;
+
+    @OneToMany(mappedBy = "member")
+    private List<Post> posts = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<PostLike> postLikes = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<Bookmark> bookmarks = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<Comment> comments = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<CommentLike> commentLikes = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<Vote> votes = new ArrayList<>();
+
+    @OneToMany(mappedBy = "member")
+    private List<Notice> notices = new ArrayList<>();
+
+    @OneToMany(mappedBy = "reporter")
+    private List<Report> reports = new ArrayList<>(); // 신고한 기록
+}
