+++ src/main/java/balancetalk/module/post/domain/BalanceOption.java
@@ -1,0 +1,49 @@
+package balancetalk.module.post.domain;
+
+import balancetalk.module.file.domain.File;
+import balancetalk.module.vote.domain.Vote;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
+import jakarta.persistence.OneToOne;
+import jakarta.validation.constraints.Size;
+import java.util.ArrayList;
+import java.util.List;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class BalanceOption {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "balance_option_id")
+    private Long id;
+
+    @NotNull
+    @Size(max = 50)
+    private String title;
+
+    @NotNull
+    @Size(max = 100)
+    private String description;
+
+    @OneToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "file_id")
+    private File file;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @OneToMany(mappedBy = "balanceOption")
+    private List<Vote> votes = new ArrayList<>();
+}
