+++ src/main/java/balancetalk/module/post/domain/PostCategory.java
@@ -1,0 +1,5 @@
+package balancetalk.module.post.domain;
+
+public enum PostCategory {
+    CASUAL, DISCUSSION
+}
