+++ src/main/java/balancetalk/module/report/domain/Report.java
@@ -1,0 +1,50 @@
+package balancetalk.module.report.domain;
+
+import balancetalk.module.comment.domain.Comment;
+import balancetalk.module.member.domain.Member;
+import balancetalk.module.post.domain.Post;
+import balancetalk.global.common.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Size;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Report extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "report_id")
+    private Long id;
+
+    @NotBlank
+    @Size(max = 300)
+    private String content;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member reporter; // 신고한 회원
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "post_id")
+    private Post post;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
+
+    @Enumerated(value = EnumType.STRING)
+    @NotNull
+    private ReportCategory category;
+}
