+++ src/main/java/balancetalk/module/vote/domain/Vote.java
@@ -1,0 +1,32 @@
+package balancetalk.module.vote.domain;
+
+import balancetalk.module.member.domain.Member;
+import balancetalk.module.post.domain.BalanceOption;
+import balancetalk.global.common.BaseTimeEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Vote extends BaseTimeEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "vote_id")
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "balance_option_id")
+    private BalanceOption balanceOption;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "member_id")
+    private Member member;
+}
